Add switch_as_x integration
