+++ homeassistant/components/switch/__init__.py
@@ -11,15 +11,12 @@ import voluptuous as vol
 from homeassistant.backports.enum import StrEnum
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import (
-    CONF_ENTITY_ID,
     SERVICE_TOGGLE,
     SERVICE_TURN_OFF,
     SERVICE_TURN_ON,
     STATE_ON,
-    Platform,
 )
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers.config_validation import (  # noqa: F401
     PLATFORM_SCHEMA,
     PLATFORM_SCHEMA_BASE,
@@ -29,7 +26,7 @@ from homeassistant.helpers.entity_component import EntityComponent
 from homeassistant.helpers.typing import ConfigType
 from homeassistant.loader import bind_hass
 
-from .const import DOMAIN
+DOMAIN = "switch"
 
 SCAN_INTERVAL = timedelta(seconds=30)
 
@@ -63,8 +60,6 @@ DEVICE_CLASSES = [cls.value for cls in SwitchDeviceClass]
 DEVICE_CLASS_OUTLET = SwitchDeviceClass.OUTLET.value
 DEVICE_CLASS_SWITCH = SwitchDeviceClass.SWITCH.value
 
-PLATFORMS: list[Platform] = [Platform.LIGHT]
-
 
 @bind_hass
 def is_on(hass: HomeAssistant, entity_id: str) -> bool:
@@ -91,21 +86,6 @@ async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up a config entry."""
-    if entry.domain == DOMAIN:
-        registry = er.async_get(hass)
-        try:
-            er.async_validate_entity_id(registry, entry.options[CONF_ENTITY_ID])
-        except vol.Invalid:
-            # The entity is identified by an unknown entity registry ID
-            _LOGGER.error(
-                "Failed to setup light switch for unknown entity %s",
-                entry.options[CONF_ENTITY_ID],
-            )
-            return False
-
-        hass.config_entries.async_setup_platforms(entry, PLATFORMS)
-        return True
-
     component: EntityComponent = hass.data[DOMAIN]
     return await component.async_setup_entry(entry)
