--- homeassistant/components/switch/const.py
@@ -1,3 +1,0 @@
-"""Constants for the Switch integration."""
-
-DOMAIN = "switch"
