+++ homeassistant/components/switch_as_x/config_flow.py
@@ -1,4 +1,4 @@
-"""Config flow for Switch integration."""
+"""Config flow for Switch as X integration."""
 from __future__ import annotations
 
 from collections.abc import Mapping
@@ -13,7 +13,7 @@ from homeassistant.helpers import (
     selector,
 )
 
-from .const import DOMAIN
+from . import DOMAIN
 
 CONFIG_FLOW = {
     "user": helper_config_entry_flow.HelperFlowStep(
@@ -31,7 +31,7 @@ CONFIG_FLOW = {
 class SwitchLightConfigFlowHandler(
     helper_config_entry_flow.HelperConfigFlowHandler, domain=DOMAIN
 ):
-    """Handle a config or options flow for Switch Light."""
+    """Handle a config flow for Switch as X."""
 
     config_flow = CONFIG_FLOW
