+++ homeassistant/components/switch_as_x/light.py
@@ -1,0 +1,32 @@
+"""Light support for switch entities."""
+from __future__ import annotations
+
+from homeassistant.components.switch.light import LightSwitch
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.const import CONF_ENTITY_ID
+from homeassistant.core import HomeAssistant
+from homeassistant.helpers import entity_registry as er
+from homeassistant.helpers.entity_platform import AddEntitiesCallback
+
+
+async def async_setup_entry(
+    hass: HomeAssistant,
+    config_entry: ConfigEntry,
+    async_add_entities: AddEntitiesCallback,
+) -> None:
+    """Initialize Light Switch config entry."""
+
+    registry = er.async_get(hass)
+    entity_id = er.async_validate_entity_id(
+        registry, config_entry.options[CONF_ENTITY_ID]
+    )
+
+    async_add_entities(
+        [
+            LightSwitch(
+                config_entry.title,
+                entity_id,
+                config_entry.entry_id,
+            )
+        ]
+    )
