+++ homeassistant/components/switch_as_x/manifest.json
@@ -1,0 +1,12 @@
+{
+  "domain": "switch_as_x",
+  "name": "Switch as X",
+  "documentation": "https://www.home-assistant.io/integrations/switch_as_x",
+  "dependencies": ["switch"],
+  "codeowners": [
+    "@home-assistant/core"
+  ],
+  "quality_scale": "internal",
+  "iot_class": "local_push",
+  "config_flow": true
+}
