+++ tests/components/switch/test_light.py
@@ -5,12 +5,8 @@ from homeassistant.components.light import (
     ATTR_SUPPORTED_COLOR_MODES,
     COLOR_MODE_ONOFF,
 )
-from homeassistant.components.switch.const import DOMAIN as SWITCH_DOMAIN
-from homeassistant.core import HomeAssistant
-from homeassistant.helpers import entity_registry as er
 from homeassistant.setup import async_setup_component
 
-from tests.common import MockConfigEntry
 from tests.components.light import common
 from tests.components.switch import common as switch_common
 
@@ -100,69 +96,3 @@ async def test_switch_service_calls(hass):
 
     assert hass.states.get("switch.decorative_lights").state == "on"
     assert hass.states.get("light.light_switch").state == "on"
-
-
-async def test_config_entry(hass: HomeAssistant):
-    """Test light switch setup from config entry."""
-    config_entry = MockConfigEntry(
-        data={},
-        domain=SWITCH_DOMAIN,
-        options={"entity_id": "switch.abc"},
-        title="ABC",
-    )
-
-    config_entry.add_to_hass(hass)
-
-    assert await hass.config_entries.async_setup(config_entry.entry_id)
-    await hass.async_block_till_done()
-
-    assert SWITCH_DOMAIN in hass.config.components
-
-    state = hass.states.get("light.abc")
-    assert state.state == "unavailable"
-    # Name copied from config entry title
-    assert state.name == "ABC"
-
-    # Check the light is added to the entity registry
-    registry = er.async_get(hass)
-    entity_entry = registry.async_get("light.abc")
-    assert entity_entry.unique_id == config_entry.entry_id
-
-
-async def test_config_entry_uuid(hass: HomeAssistant):
-    """Test light switch setup from config entry with entity registry id."""
-    registry = er.async_get(hass)
-    registry_entry = registry.async_get_or_create("switch", "test", "unique")
-
-    config_entry = MockConfigEntry(
-        data={},
-        domain=SWITCH_DOMAIN,
-        options={"entity_id": registry_entry.id},
-        title="ABC",
-    )
-
-    config_entry.add_to_hass(hass)
-
-    assert await hass.config_entries.async_setup(config_entry.entry_id)
-    await hass.async_block_till_done()
-
-    assert hass.states.get("light.abc")
-
-
-async def test_config_entry_unregistered_uuid(hass: HomeAssistant):
-    """Test light switch setup from config entry with unknown entity registry id."""
-    fake_uuid = "a266a680b608c32770e6c45bfe6b8411"
-
-    config_entry = MockConfigEntry(
-        data={},
-        domain=SWITCH_DOMAIN,
-        options={"entity_id": fake_uuid},
-        title="ABC",
-    )
-
-    config_entry.add_to_hass(hass)
-
-    assert not await hass.config_entries.async_setup(config_entry.entry_id)
-    await hass.async_block_till_done()
-
-    assert len(hass.states.async_all()) == 0
