+++ tests/components/switch_as_x/__init__.py
@@ -1,0 +1,1 @@
+"""The tests for Switch as X platforms."""
